refactor: move didcomm+protocol handlers to separate layer
