+++ agents/rust/mediator/src/didcomm_handlers/utils.rs
@@ -1,0 +1,9 @@
+pub mod prelude {
+    pub use std::sync::Arc;
+
+    pub use aries_vcx::utils::encryption_envelope::EncryptionEnvelope;
+    pub use aries_vcx_core::wallet::base_wallet::BaseWallet;
+    pub use mediation::storage::MediatorPersistence;
+
+    pub use crate::{aries_agent::ArcAgent, utils::prelude::*};
+}
