+++ agents/rust/mediator/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod aries_agent;
+pub mod didcomm_handlers;
 pub mod http_routes;
 pub mod utils;
